rename: add prefix 'header' to header buttons
