+++ src/widgets/LayoutHeader/ui/HeaderCartButton/HeaderCartButton.module.css
@@ -1,4 +1,4 @@
-.cartButtonContainer {
+.headerCartButtonContainer {
   position: relative;
   width: 32px;
   height: 32px;
