+++ src/widgets/LayoutHeader/ui/HeaderLogoButton/HeaderLogoButton.module.css
@@ -1,4 +1,4 @@
-.logoButton {
+.headerLogoButton {
   border: none;
   height: 16px;
 }
